requirements.txt
pandas
numpy 
scikit-learn
requests
bentoml
pydantic
pyjwt
pytest